+++ src/vector/1_vec.c
@@ -6,13 +6,13 @@
 /*   By: mkulikov <mkulikov@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 10:54:53 by mkulikov          #+#    #+#             */
-/*   Updated: 2024/10/18 12:48:18 by mkulikov         ###   ########.fr       */
+/*   Updated: 2024/10/21 15:09:40 by mkulikov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/minirt.h"
 
-t_vector	*new_vec(double x, double y, double z)
+t_vector	*new_vec(float x, float y, float z)
 {
 	t_vector	*v;
 
@@ -27,9 +27,9 @@ t_vector	*new_vec(double x, double y, double z)
 
 t_vector	*vec_sub(t_vector *v1, t_vector *v2)
 {
-	double		x;
-	double		y;
-	double		z;
+	float		x;
+	float		y;
+	float		z;
 	t_vector	*v;
 
 	x = v1->x - v2->x;
@@ -39,9 +39,9 @@ t_vector	*vec_sub(t_vector *v1, t_vector *v2)
 	return (v);
 }
 
-double	vec_len(t_vector *v)
+float	vec_len(t_vector *v)
 {
-	double	len;
+	float	len;
 
 	len = sqrt((v->x * v->x) + (v->y * v->y) + (v->z * v->z));
 	return (len);
@@ -49,7 +49,7 @@ double	vec_len(t_vector *v)
 
 void		vec_norm(t_vector *v)
 {
-	double	len;
+	float	len;
 
 	len = vec_len(v);
 	v->x /= len;
@@ -57,9 +57,9 @@ void		vec_norm(t_vector *v)
 	v->z /= len;
 }
 
-double	vec_dot_prod(t_vector *v1, t_vector *v2)
+float	vec_dot_prod(t_vector *v1, t_vector *v2)
 {
-	double	dot_prod;
+	float	dot_prod;
 
 	dot_prod = (v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z);
 	return (dot_prod);
